#369 Move the constraint method to specific diagram builder
